added sally_description URDF package into git
