--- README.md
@@ -1,2 +1,0 @@
-# wall-climber
-Codebase for the Sally magnetic wall-climbing robot
