--- package.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>wall_climber</name>
-  <version>0.0.0</version>
-  <description>Codebase for the Sally magnetic wall-climbing robot</description>
-  <maintainer email="pnadan@cmu.edu">Paul Nadan</maintainer>
-  <maintainer email="nrsun@andrew.cmu.edu">Nathan Sun</maintainer>
-  <license>MIT</license>
-
-  <exec_depend>rclpy</exec_depend>
-  <exec_depend>std_msgs</exec_depend>
-  <exec_depend>sensor_msgs</exec_depend>
-  <exec_depend>microstrain_inertial_driver</exec_depend>
-
-  <test_depend>ament_copyright</test_depend>
-  <test_depend>ament_flake8</test_depend>
-  <test_depend>ament_pep257</test_depend>
-  <test_depend>python3-pytest</test_depend>
-
-  <export>
-    <build_type>ament_python</build_type>
-  </export>
-</package>
