--- requirements.txt
@@ -1,4 +1,0 @@
-dynamixel_sdk
-numpy
-pyserial
-inputs
