--- setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/wall_climber
-[install]
-install_scripts=$base/lib/wall_climber
