+++ wall_climber/README.md
@@ -1,0 +1,2 @@
+# wall-climber
+Codebase for the Sally magnetic wall-climbing robot
