+++ wall_climber/launch/wc_pxrf.launch
@@ -1,0 +1,18 @@
+<launch>
+
+    <node pkg="joy" exec="joy_node" name="joy_node" respawn="true">
+        <param name="dev" value="/dev/input/js0" />
+    </node>
+
+    <node name="main" exec="main.py" pkg="wall-climber" output="screen"/>
+
+    <node name="gui" exec="gui.py" pkg="pxrf" output="screen"/>
+
+     <!-- Launch the Microstrain node so the IMU starts publishing data -->
+    <node name="microstrain_intertial_driver" pkg="microstrain_inertial_driver" exec="microstrain_inertial_driver_node" output="screen" namespace="/">
+
+        <!-- Load the default params file. -->
+        <param from="$(find microstrain_inertial_driver)/microstrain_inertial_driver_common/config/params.yml"/>
+    </node>
+
+</launch>
