+++ wall_climber/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/wall_climber
+[install]
+install_scripts=$base/lib/wall_climber
